fixed compile errors and warnings using Dev-C++ in "Compile as C++" mode
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@9198 d7cf8633-e32d-0410-b094-e92efae38249

+++ cap_ogg.c
@@ -1009,13 +1009,13 @@ void SCR_CaptureVideo_Ogg_BeginVideo(void)
 			if(ti.quality < 0)
 			{
 				ti.target_bitrate = -1;
-				ti.keyframe_data_target_bitrate = -1;
+				ti.keyframe_data_target_bitrate = (unsigned int)-1;
 				ti.quality = 63;
 			}
 			else
 			{
 				ti.target_bitrate = -1;
-				ti.keyframe_data_target_bitrate = -1;
+				ti.keyframe_data_target_bitrate = (unsigned int)-1;
 				ti.quality = bound(0, ti.quality, 63);
 			}
 		}
